🔨 refactor(app): moved files around to make poject script work
